Drive generated output for double

+++ tests/fixtures/migrations/model-modifiers.php
@@ -19,7 +19,7 @@ class CreateModifiersTable extends Migration
             $table->string('name', 1000)->unique()->charset('utf8');
             $table->string('content')->default('');
             $table->float('amount', 9, 3);
-            $table->decimal('total', 10, 2);
+            $table->double('total', 10, 2);
             $table->decimal('overflow', 99, 99);
             $table->char('ssn', 11);
             $table->enum('role', ["user","admin","owner"]);
